+++ torch_geometric/nn/pool/connect/__init__.py
@@ -1,0 +1,3 @@
+from .base import Connect
+
+__all__ = ['Connect']
