+++ torch_geometric/nn/pool/select/__init__.py
@@ -1,0 +1,3 @@
+from .base import Select
+
+__all__ = ['Select']
